A great many changes Namely, this commit removes `SliceTokenizer`, and effectively adds the API as an extension directly to `Tokenizer`. In doing this, I added a comptime flag to the implementation to elide the branch which checks for whether to return `error.BufferUnderrun`, when we know from the callsite whether or not that is possible. This culminates in the `*AssumeNoUnderrun` API which, as in `SliceTokenizer`, does not return `error.BufferUnderrun`.
During this, I also reworked the implementation to merge `nextTypeImpl`
and `nextSrcImpl` into `nextTypeOrSrcImpl`, which brings the branches
for the token type and token source right next to each other - this
should have no effect on code generation, it is a purely organisational
change.

I also went through the code after the fact and removed `inline`
branches wherever practical, to reduce the generated code size.

This also adds (back) `nextSrc` to `ReadingTokenizer`, with test
coverage to confirm it works as I think it's supposed to.